refactor: FavoriteTeam 관련 UI 컴포넌트 위치 이동
