+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/navigation/NavigationRoot.kt
@@ -17,7 +17,7 @@ import androidx.navigation3.runtime.NavKey
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.ui.NavDisplay
 import com.yagubogu.ui.badge.component.BadgeScreen
-import com.yagubogu.ui.favorite.FavoriteTeamScreen
+import com.yagubogu.ui.onboarding.favorite.FavoriteTeamScreen
 import com.yagubogu.ui.livetalk.chat.LivetalkChatScreen
 import com.yagubogu.ui.login.LoginScreen
 import com.yagubogu.ui.main.MainScreen
